authorization

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,31 +14,31 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getAll(filterDto: GetAllFilterDto): Promise<Task[]> {
-    return this.taskRepository.getAll(filterDto);
+  async getAll(filterDto: GetAllFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getAll(filterDto, user);
   }
 
-  async getById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!found) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
     return found;
   }
 
-  async create(dto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createEntity(dto);
+  async create(dto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createEntity(dto, user);
   }
 
-  async updateStatusById(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getById(id);
+  async updateStatusById(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async deleteById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
